FEM WIP Lagrange unit test works, ignoring ghost cells in assembly
